chore: move stage names up one level
